move plot_fits_image function to its own file fits_image.py
